src/requirements.txt
fastapi<0.110
uvicorn[standard]
python-socketio
aiortc
opencv-python
numpy
pydantic
jinja2
aiofiles
python-multipart
requests
pyaudio
python-vlc
psutil
python-jose[cryptography]
httpx[http2]
speedtest-cli
sqlalchemy
audioop-lts
passlib
argon2_cffi
colorlog